# openapi/vault-admin-api.yaml
openapi: 3.0.3
info:
  title: Vault Admin API
  description: |
    **Administrative API for Vault** - User and system management for administrators.
    
    This API is designed for admin dashboards and management tools. It provides:
    - Full user management (CRUD, suspend, audit)
    - Organization management (as admin)
    - Tenant configuration
    - Audit logs access
    - System settings
    
    ## Authentication
    
    All endpoints require:
    1. JWT Bearer token (same as Client API)
    2. Admin role (`admin` or `owner`)
    
    ```
    Authorization: Bearer <jwt-token>
    ```
    
    ## Authorization
    
    Endpoints have role-based access:
    - `owner` - Full access to everything
    - `admin` - Can manage users and orgs, cannot delete tenant
    
    ## Multi-Tenancy
    
    Admin requests are scoped to a tenant. Admins can only manage resources within their tenant.
  version: 0.1.0
  contact:
    name: Vault Team
  license:
    name: MIT OR Apache-2.0
servers:
  - url: https://api.vault.dev/api/v1/admin
    description: Production
  - url: http://localhost:3000/api/v1/admin
    description: Local development
...
paths:
...
  /users/{userId}:
    get:
      summary: Get user details
      operationId: getUser
      tags: [User Management]
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      summary: Update user
      operationId: updateUser
      tags: [User Management]
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete user (hard delete)
      description: Permanently delete a user account
      operationId: deleteUser
      tags: [User Management]
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /settings/security:
    get:
      summary: Get security settings (v2)
      operationId: getSecuritySettings
      tags: [Tenant Settings]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Security settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecuritySettingsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    patch:
      summary: Update security settings (v2)
      operationId: updateSecuritySettings
      tags: [Tenant Settings]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecuritySettings'
      responses:
        '200':
          description: Security settings updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecuritySettings'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Admin permission required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object }
    MessageResponse:
      type: object
      properties:
        message: { type: string }
...
    UpdateUserRequest:
      type: object
      properties:
        email: { type: string, format: email }
        name: { type: string }
        status: { type: string, enum: [active, pending, suspended, deactivated] }
    AdminUserResponse:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        email_verified: { type: boolean }
        name: { type: string, nullable: true }
        status: { type: string }
        last_login_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        organization_count: { type: integer }
...
    SecuritySettingsResponse:
      type: object
      properties:
        tenant_id: { type: string }
        category: { type: string }
        settings:
          $ref: '#/components/schemas/SecuritySettings'
        updated_at: { type: string, format: date-time }
...
    SecuritySettings:
      type: object
      properties:
        password_policy:
          type: object
          properties:
            min_length: { type: integer }
            max_length: { type: integer }
            require_uppercase: { type: boolean }
            require_lowercase: { type: boolean }
            require_numbers: { type: boolean }
            require_special: { type: boolean }
            special_chars: { type: string }
            max_consecutive_chars: { type: integer }
            prevent_common_passwords: { type: boolean }
            history_count: { type: integer }
            expiry_days: { type: integer, nullable: true }
            check_breach: { type: boolean }
            enforcement_mode: { type: string, enum: [block, warn, audit] }
            min_entropy: { type: number }
            prevent_user_info: { type: boolean }
        session_lifetime:
          type: object
          properties:
            access_token_minutes: { type: integer }
            refresh_token_days: { type: integer }
            absolute_timeout_hours: { type: integer }
            idle_timeout_minutes: { type: integer }
        session_limits:
          type: object
          properties:
            max_concurrent_sessions: { type: integer }
            eviction_policy: { type: string }
            enforce_for_ip: { type: boolean }
            max_sessions_per_ip: { type: integer }
        mfa_settings:
          type: object
          properties:
            require_mfa: { type: boolean }
            allowed_methods:
              type: array
              items: { type: string, enum: [totp, email, sms, webauthn] }
            grace_period_days: { type: integer }
            require_mfa_for_roles:
              type: array
              items: { type: string }
        lockout_policy:
          type: object
          properties:
            max_failed_attempts: { type: integer }
            lockout_duration_minutes: { type: integer }
            reset_after_minutes: { type: integer }
        notifications:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/SecurityNotificationAudience'
            admin:
              $ref: '#/components/schemas/SecurityNotificationAudience'
            admin_roles:
              type: array
              items: { type: string }
            whatsapp_template_name:
              type: string
              nullable: true
    SecurityNotificationAudience:
      type: object
      properties:
        enabled: { type: boolean }
        events:
          type: array
          items:
            type: string
            enum:
              - login_failed
              - login_blocked_risk
              - password_changed
              - password_reset
              - mfa_enabled
              - mfa_disabled
              - suspicious_login
              - account_locked
              - impersonation_started
              - security_policy_updated
        channels:
          type: array
          items:
            type: string
            enum: [email, sms, whatsapp]
...